update from original

+++ src/engine.rs
@@ -1,5 +1,3 @@
-extern crate stopwatch;
-
 use std::rc::Rc;
 use std::cell::RefCell;
 use std::thread;
@@ -9,21 +7,16 @@ use std::collections::HashMap;
 use std::intrinsics;
 use std::raw::TraitObject;
 use std::mem;
+use std::time::Duration;
 
 use bootstrap;
+use bootstrap::input::ScanCode;
 use bootstrap::window::Window;
 use bootstrap::window::Message::*;
 use bootstrap::time::Timer;
-
 use bs_audio;
-
 use polygon::gl_render::GLRender;
-
-#[cfg(feature = "timing")]
-use self::stopwatch::{Collector, Stopwatch};
-
-#[cfg(not(feature = "timing"))]
-use self::stopwatch::null::{Collector, Stopwatch};
+use stopwatch::{Collector, Stopwatch};
 
 use scene::Scene;
 use resource::ResourceManager;
@@ -40,6 +33,7 @@ pub struct Engine {
     resource_manager: Rc<ResourceManager>,
 
     systems: Vec<Box<System>>,
+    debug_systems: Vec<Box<System>>,
     system_indices: HashMap<TypeId, usize>,
     system_names: HashMap<String, TypeId>,
 
@@ -47,11 +41,14 @@ pub struct Engine {
     light_update: Box<System>,
     audio_update: Box<System>,
     alarm_update: Box<System>,
+    collision_update: Box<System>,
+
     scene: Scene,
 
     debug_draw: DebugDraw,
 
     close: bool,
+    debug_pause: bool,
 }
 
 impl Engine {
@@ -62,11 +59,9 @@ impl Engine {
         let resource_manager = Rc::new(ResourceManager::new(renderer.clone()));
 
         let audio_source = match bs_audio::init() {
-            Ok(audio_source) => {
-                println!("Audio subsystem successfully initialized");
-                audio_source
-            },
+            Ok(audio_source) => audio_source,
             Err(error) => {
+                // TODO: Rather than panicking, create a null audio system and keep running.
                 panic!("Error while initialzing audio subsystem: {}", error)
             },
         };
@@ -77,23 +72,27 @@ impl Engine {
             resource_manager: resource_manager.clone(),
 
             systems: Vec::new(),
+            debug_systems: Vec::new(),
             system_indices: HashMap::new(),
             system_names: HashMap::new(),
 
-            transform_update: Box::new(TransformUpdateSystem),
+            transform_update: Box::new(transform_update),
             light_update: Box::new(LightUpdateSystem),
             audio_update: Box::new(AudioSystem),
             alarm_update: Box::new(AlarmSystem),
+            collision_update: Box::new(CollisionSystem::new()),
+
             scene: Scene::new(&resource_manager, audio_source),
 
             debug_draw: DebugDraw::new(renderer.clone(), &*resource_manager),
 
             close: false,
+            debug_pause: false,
         }
     }
 
     pub fn update(&mut self) {
-        let _stopwatch = Stopwatch::named("update");
+        let _stopwatch = Stopwatch::new("update");
 
         let scene = &mut self.scene;
 
@@ -123,23 +122,47 @@ impl Engine {
             }
         }
 
-        self.alarm_update.update(scene, TARGET_FRAME_TIME_SECONDS);
+        // TODO: More efficient handling of debug pause (i.e. something that doesn't have any
+        // overhead when doing a release build).
+        if !self.debug_pause || scene.input.key_pressed(ScanCode::F11) {
+            self.debug_draw.clear_buffer();
 
-        // Update systems.
-        for system in self.systems.iter_mut() {
+            self.alarm_update.update(scene, TARGET_FRAME_TIME_SECONDS);
+
+            // Update systems.
+            for system in self.systems.iter_mut() {
+                system.update(scene, TARGET_FRAME_TIME_SECONDS);
+            }
+        }
+
+        // Update debug systems always forever.
+        for system in self.debug_systems.iter_mut() {
             system.update(scene, TARGET_FRAME_TIME_SECONDS);
         }
 
         self.transform_update.update(scene, TARGET_FRAME_TIME_SECONDS);
-        self.light_update.update(scene, TARGET_FRAME_TIME_SECONDS);
-        self.audio_update.update(scene, TARGET_FRAME_TIME_SECONDS);
 
-        // Cleanup any entities that have been marked for destroy.
-        scene.destroy_marked();
+        if !self.debug_pause || scene.input.key_pressed(ScanCode::F11) {
+            self.collision_update.update(scene, TARGET_FRAME_TIME_SECONDS);
+            self.light_update.update(scene, TARGET_FRAME_TIME_SECONDS);
+            self.audio_update.update(scene, TARGET_FRAME_TIME_SECONDS);
+
+            // Cleanup any entities that have been marked for destroy.
+            scene.destroy_marked();
+        }
+
+        if scene.input.key_pressed(ScanCode::F9) {
+            self.debug_pause = !self.debug_pause;
+        }
+
+        if scene.input.key_pressed(ScanCode::F11) {
+            self.debug_pause = true;
+        }
     }
 
+    #[cfg(not(feature="no-draw"))]
     pub fn draw(&mut self) {
-        let _stopwatch = Stopwatch::named("draw");
+        let _stopwatch = Stopwatch::new("draw");
 
         self.renderer.clear();
 
@@ -179,12 +202,15 @@ impl Engine {
         self.renderer.swap_buffers(self.window.borrow().deref());
     }
 
+    #[cfg(feature="no-draw")]
+    pub fn draw(&mut self) {}
+
     pub fn main_loop(&mut self) {
         let timer = Timer::new();
         let mut collector = Collector::new().unwrap();
 
         loop {
-            let _stopwatch = Stopwatch::named("loop");
+            let _stopwatch = Stopwatch::new("loop");
 
             let start_time = timer.now();
 
@@ -195,14 +221,15 @@ impl Engine {
                 break;
             }
 
-            if timer.elapsed_ms(start_time) > TARGET_FRAME_TIME_MS {
+            if !cfg!(feature="timing")
+            && timer.elapsed_ms(start_time) > TARGET_FRAME_TIME_MS {
                 println!("WARNING: Missed frame time. Frame time: {}ms, target frame time: {}ms", timer.elapsed_ms(start_time), TARGET_FRAME_TIME_MS);
             }
 
             // Wait for target frame time.
             let mut remaining_time_ms = TARGET_FRAME_TIME_MS - timer.elapsed_ms(start_time);
             while remaining_time_ms > 1.0 {
-                thread::sleep_ms(remaining_time_ms as u32);
+                thread::sleep(Duration::from_millis(remaining_time_ms as u64));
                 remaining_time_ms = TARGET_FRAME_TIME_MS - timer.elapsed_ms(start_time);
             }
 
@@ -225,6 +252,10 @@ impl Engine {
         self.systems.push(Box::new(system));
         self.system_indices.insert(system_id, index);
         self.system_names.insert(type_name::<T>().into(), system_id);
+    }
+
+    pub fn register_debug_system<T: Any + System>(&mut self, system: T) {
+        self.debug_systems.push(Box::new(system));
     }
 
     pub fn get_system<T: Any + System>(&self) -> &T {
@@ -282,18 +313,22 @@ impl Clone for Engine {
             resource_manager: resource_manager.clone(),
 
             systems: Vec::new(),
+            debug_systems: Vec::new(),
             system_indices: HashMap::new(),
             system_names: HashMap::new(),
 
-            transform_update: Box::new(TransformUpdateSystem),
+            transform_update: Box::new(transform_update),
             light_update: Box::new(LightUpdateSystem),
             audio_update: Box::new(AudioSystem),
             alarm_update: Box::new(AlarmSystem),
+            collision_update: Box::new(CollisionSystem::new()),
+
             scene: self.scene.clone(&resource_manager),
 
             debug_draw: DebugDraw::new(self.renderer.clone(), &*resource_manager),
 
             close: false,
+            debug_pause: false,
         };
 
         engine
@@ -327,18 +362,22 @@ pub fn engine_init(window: Rc<RefCell<Window>>) -> Box<Engine> {
         resource_manager: resource_manager.clone(),
 
         systems: Vec::new(),
+        debug_systems: Vec::new(),
         system_indices: HashMap::new(),
         system_names: HashMap::new(),
 
-        transform_update: Box::new(TransformUpdateSystem),
+        transform_update: Box::new(transform_update),
         light_update: Box::new(LightUpdateSystem),
         audio_update: Box::new(AudioSystem),
         alarm_update: Box::new(AlarmSystem),
+        collision_update: Box::new(CollisionSystem::new()),
+
         scene: Scene::new(&resource_manager, audio_source),
 
         debug_draw: DebugDraw::new(renderer.clone(), &*resource_manager),
 
         close: false,
+        debug_pause: false,
     })
 }
 
@@ -362,4 +401,11 @@ pub fn engine_close(engine: &Engine) -> bool {
 #[no_mangle]
 pub fn engine_drop(engine: Box<Engine>) {
     drop(engine);
+}
+
+// #[cfg(test)] // TODO: Only include this for benchmarks. Double TODO: better support headless benches.
+pub fn do_collision_update(engine: &mut Engine) {
+    let scene = &mut engine.scene;
+    engine.transform_update.update(scene, TARGET_FRAME_TIME_SECONDS);
+    engine.collision_update.update(scene, TARGET_FRAME_TIME_SECONDS);
 }
